+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+- Added support for the `stm32f3xx-hal` crate.
+
 ## [0.1.0-alpha.1] - 2024-07-06
 
 [v0.1.0-alpha.1]: https://github.com/sunsided/chip-select/releases/tag/v0.1.0-alpha.1